friendlier histogram intervals with condensing
Each candidate histogram in the time histogram now has nice human-friendly
intervals.  Sometimes this means that in the worst case there may be so many
bars as to be unreadable (a day broken up into 10 second intervals).  In this
case the bars are condensed to a more readable size before rendering