tcpflow -c should print a dot for non-printable chars
Older versions of tcpflow would print a dot for each non-printable character when running "tcpflow -c".  The manpage confirms that this is the expected behavior.

The current version doesn't print the dot (or any other character) when encountering non-printable characters.

This quick and dirty Pull Request adds the else statement to print the dot.