Tweak example build commands (#140)
bootstrap.sh will fail on most systems if run with "sh bootstrap.sh"
because it requires Bash and the system shell is usually something
faster. It will succeed if run with "./boostrap.sh" because it specifies
Bash with a shebang.

We might as well change "sh configure" to "./configure" as well because
the latter is more reliable and idiomatic.