Clean shutdown of tcpflow
The standard behavior of tcpflow when receiving a terminating signal
(such as SIGINT with Ctrl-C) is to print 'terminating' and do a quick
exit which will not properly close the capture files.

This changes the shutdown logic so that pcap_breakloop(..) is called
in the signal handler for SIGINT and SIGHUP, therefore ensuring a
proper shutdown that closes all files.